add right sidebar, autohide sidebars
the aim is not to show sidebars on phones and such. Usually it doesn't
make a foog user experience to show them above or below the content.
There should be a dedicated page, like archive where you can browse
easily older posts, and as for categories, I don't know yet...